Renaming precheck services
Create a precheck namespace and put precheck related services inside.
