+++ app/controllers/precheck/confirmation_controller.rb
@@ -1,8 +1,8 @@
 class Precheck::ConfirmationController < PrecheckController
   def create
-    redirect_to precheck_status_path and return unless PrecheckEligibilityService.new(family: @family).call
+    redirect_to precheck_status_path and return unless Precheck::EligibilityService.new(family: @family).call
 
     @family.update!(precheck_status: :approved)
-    UpdatedFamilyPrecheckStatusService.new(family: @family).call
+    Precheck::UpdatedFamilyStatusService.new(family: @family).call
   end
 end
