+++ app/controllers/precheck/rejection_controller.rb
@@ -1,9 +1,9 @@
 class Precheck::RejectionController < PrecheckController
   def create
-    redirect_to precheck_status_path and return if PrecheckEligibilityService.new(family: @family).too_late_or_checked_in?
+    redirect_to precheck_status_path and return if Precheck::EligibilityService.new(family: @family).too_late_or_checked_in?
 
     @family.update!(precheck_status: :changes_requested)
     @message = params['message']
-    UpdatedFamilyPrecheckStatusService.new(family: @family, message: @message).call
+    Precheck::UpdatedFamilyStatusService.new(family: @family, message: @message).call
   end
 end
