+++ app/jobs/nightly_precheck_mailer_job.rb
@@ -3,15 +3,21 @@ class NightlyPrecheckMailerJob < ApplicationJob
 
   def perform
     families_scope.find_each do |family|
-      if precheck_eligible?(family)
-        PrecheckMailer.confirm_charges(family).deliver_later
-      elsif actionable_errors?(family)
-        PrecheckMailer.report_issues(family).deliver_later
-      end
+      deliver_pending_approval_mail(family)
     end
   end
 
   private
+
+  def deliver_pending_approval_mail(family)
+    return unless family.pending_approval?
+
+    if precheck_eligible?(family)
+      PrecheckMailer.confirm_charges(family).deliver_later
+    elsif actionable_errors?(family)
+      PrecheckMailer.report_issues(family).deliver_later
+    end
+  end
 
   def families_scope
     Family.includes(
@@ -20,7 +26,7 @@ class NightlyPrecheckMailerJob < ApplicationJob
   end
 
   def eligibility_service(family)
-    PrecheckEligibilityService.new(family: family)
+    Precheck::EligibilityService.new(family: family)
   end
 
   def precheck_eligible?(family)
